+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 master
 ------
 
+* Generate `DS.hasMany` and `DS.belongsTo` relationships from JSON Schema.
+
 0.0.1
 -----